Testing git functionality + various cleanup
Refactored courses.views.regional_earnings() to be more DRY by using a for loop with three iterations rather than three blocks of similar code. Reconfigured json creation process in this function to pull from a dictionary that is gradually filled with values during each loop. Also moved the 'unhandled error' json response to the top of the function to act as a guard, avoiding unnecessary indentation.